+++ src/views/network/AntennaCriteriaCheck.vue
@@ -4,6 +4,10 @@
       <article class="tile is-child">
         <h4 class="title">Antenna Criteria Check</h4>
 
+        <div class="notification is-warning">
+          The computation of the fulfilment of 'Events' is currently not working correctly. It does <strong>not</strong> yet take into account the Summer Universities organised. Please check these manually!
+        </div>
+
         <div class="field">
           <label class="label">Selected Agora</label>
           <div class="select">
@@ -44,22 +48,22 @@
             </b-table-column>
 
             <b-table-column field="boardElection" label="Board election (BE)">
-              <b-tag type="is-success" size="is-medium" v-if="props.row.antennaCriteria.boardElection && !showDetails">Yes</b-tag>
-              <b-tag type="is-success" size="is-medium" v-if="props.row.antennaCriteria.boardElection && showDetails">
+              <b-tag type="is-success" size="is-medium" v-if="props.row.antennaCriteria.boardElection === 'true' && !showDetails">Yes</b-tag>
+              <b-tag type="is-success" size="is-medium" v-if="props.row.antennaCriteria.boardElection === 'true' && showDetails">
                 {{ props.row.latest_election }}
               </b-tag>
-              <b-tag type="is-danger" size="is-medium" v-if="!props.row.antennaCriteria.boardElection && !showDetails && props.row.type === 'antenna'">No</b-tag>
-              <b-tag type="is-danger" size="is-medium" v-if="!props.row.antennaCriteria.boardElection && showDetails && props.row.latest_election && props.row.type === 'antenna'">
+              <b-tag type="is-danger" size="is-medium" v-if="props.row.antennaCriteria.boardElection !== 'true' && !showDetails && props.row.type === 'antenna'">No</b-tag>
+              <b-tag type="is-danger" size="is-medium" v-if="props.row.antennaCriteria.boardElection !== 'true' && showDetails && props.row.latest_election && props.row.type === 'antenna'">
                 {{ props.row.latest_election }}
               </b-tag>
-              <b-tag type="is-danger" size="is-medium" v-if="!props.row.antennaCriteria.boardElection && showDetails && !props.row.latest_election && props.row.type === 'antenna'">No</b-tag>
-              <b-tag type="is-info" size="is-medium" v-if="!props.row.antennaCriteria.boardElection && props.row.type !== 'antenna'">Else</b-tag>
+              <b-tag type="is-danger" size="is-medium" v-if="props.row.antennaCriteria.boardElection !== 'true' && showDetails && !props.row.latest_election && props.row.type === 'antenna'">No</b-tag>
+              <b-tag type="is-info" size="is-medium" v-if="props.row.antennaCriteria.boardElection !== 'true' && props.row.type !== 'antenna'">Else</b-tag>
             </b-table-column>
 
             <b-table-column field="membersList" label="Members list (ML)">
-              <b-tag type="is-success" size="is-medium" v-if="props.row.antennaCriteria.membersList">Yes</b-tag>
-              <b-tag type="is-danger" size="is-medium" v-if="!props.row.antennaCriteria.membersList && props.row.type !== 'contact'">No</b-tag>
-              <b-tag type="is-info" size="is-medium" v-if="!props.row.antennaCriteria.membersList && props.row.type === 'contact'">Else</b-tag>
+              <b-tag type="is-success" size="is-medium" v-if="props.row.antennaCriteria.membersList === 'true'">Yes</b-tag>
+              <b-tag type="is-danger" size="is-medium" v-if="props.row.antennaCriteria.membersList !== 'true' && props.row.type !== 'contact'">No</b-tag>
+              <b-tag type="is-info" size="is-medium" v-if="props.row.antennaCriteria.membersList !== 'true' && props.row.type === 'contact'">Else</b-tag>
             </b-table-column>
 
             <b-table-column field="membershipFee" label="Membership fee (F)">
@@ -75,16 +79,16 @@
             </b-table-column>
 
             <b-table-column field="mostRecentEvent" label="Events (E)">
-              <b-tag type="is-success" size="is-medium" v-if="props.row.antennaCriteria.events && !showDetails">Yes</b-tag>
-              <b-tag type="is-success" size="is-medium" v-if="props.row.antennaCriteria.events && showDetails">
+              <b-tag type="is-success" size="is-medium" v-if="props.row.antennaCriteria.events === 'true' && !showDetails">Yes</b-tag>
+              <b-tag type="is-success" size="is-medium" v-if="props.row.antennaCriteria.events === 'true' && showDetails">
                 {{ props.row.latest_event }}
               </b-tag>
-              <b-tag type="is-danger" size="is-medium" v-if="!props.row.antennaCriteria.events && !showDetails && props.row.type === 'antenna'">No</b-tag>
-              <b-tag type="is-danger" size="is-medium" v-if="!props.row.antennaCriteria.events && showDetails && props.row.latest_event && props.row.type === 'antenna'">
+              <b-tag type="is-danger" size="is-medium" v-if="props.row.antennaCriteria.events !== 'true' && !showDetails && props.row.type === 'antenna'">No</b-tag>
+              <b-tag type="is-danger" size="is-medium" v-if="props.row.antennaCriteria.events !== 'true' && showDetails && props.row.latest_event && props.row.type === 'antenna'">
                 {{ props.row.latest_event }}
               </b-tag>
-              <b-tag type="is-danger" size="is-medium" v-if="!props.row.antennaCriteria.events && showDetails && !props.row.latest_event && props.row.type === 'antenna'">No</b-tag>
-              <b-tag type="is-info" size="is-medium" v-if="!props.row.antennaCriteria.events && props.row.type !== 'antenna'">Else</b-tag>
+              <b-tag type="is-danger" size="is-medium" v-if="props.row.antennaCriteria.events !== 'true' && showDetails && !props.row.latest_event && props.row.type === 'antenna'">No</b-tag>
+              <b-tag type="is-info" size="is-medium" v-if="props.row.antennaCriteria.events !== 'true' && props.row.type !== 'antenna'">Else</b-tag>
             </b-table-column>
 
             <b-table-column field="attendance" label="Agora attendance (AA)">
@@ -164,7 +168,8 @@ export default {
       events: [],
       statutoryEvents: [],
       summerUniversities: [],
-      isLoading: false
+      isLoading: false,
+      isLoadingAgora: false
     }
   },
   computed: {
@@ -213,7 +218,7 @@ export default {
       this.hideSafeLocals = !this.hideSafeLocals
     },
     fetchAgorae () {
-      this.isLoading = true
+      this.isLoadingAgora = true
       this.axios.get(this.services['statutory'], { params: { type: 'agora' } }).then((response) => {
         this.agorae = response.data.data
         if (this.agorae.length === 0) {
@@ -222,9 +227,9 @@ export default {
         // Automatically set the most recent Agora as the selected one
         this.selectedAgora = this.agorae[0]
         this.fetchData()
-        this.isLoading = false
+        this.isLoadingAgora = false
       }).catch((err) => {
-        this.isLoading = false
+        this.isLoadingAgora = false
         this.$root.showError('Could not fetch statutory data', err)
       })
     },
@@ -234,14 +239,18 @@ export default {
       this.axios.get(this.services['core'] + '/bodies').then(async (bodiesResponse) => {
         this.bodies = bodiesResponse.data.data
         this.bodies = this.bodies.filter(x => ['antenna', 'contact antenna', 'contact'].includes(x.type))
-        this.bodies.forEach(body => { body.antennaCriteria = {} })
+        this.bodies.forEach(body => {
+          body.antennaCriteria = {}
+          body.comments = {}
+        })
 
         const promises = []
         promises.push(this.checkBoardCriterium())
         promises.push(this.checkMembersList())
         promises.push(this.checkEventsCriterium())
 
-        await Promise.all(promises)
+        // The allSettled() command waits for all promises to be done, so it is also 'fine' if some of them fail
+        await Promise.allSettled(promises)
 
         // Do this after the rest, to make sure it also "overrides" automatically computed fields
         await this.getAntennaCriteriaFulfilment()
@@ -279,26 +288,23 @@ export default {
       })
     },
     async fetchEvents () {
-      await this.axios.get(this.services['events'] + '/recents', { params: { ends: this.selectedAgora.ends } }).then((response) => {
-        this.events = response.data.data
+      await this.axios.get(this.services['events'] + '/recents', { params: { ends: this.selectedAgora.ends } }).then((eventResponse) => {
+        this.events = eventResponse.data.data
       }).catch((err) => {
-        this.isLoading = false
         this.$root.showError('Could not fetch event data', err)
       })
     },
     async fetchStatutoryEvents () {
-      await this.axios.get(this.services['statutory'] + '/recents', { params: { ends: this.selectedAgora.ends } }).then((response) => {
-        this.statutoryEvents = response.data.data
+      await this.axios.get(this.services['statutory'] + '/recents', { params: { ends: this.selectedAgora.ends } }).then((statutoryResponse) => {
+        this.statutoryEvents = statutoryResponse.data.data
       }).catch((err) => {
-        this.isLoading = false
         this.$root.showError('Could not fetch statutory event data', err)
       })
     },
     async fetchSummerUniversities () {
-      await this.axios.get(this.services['summeruniversity'] + '/recents', { params: { ends: this.selectedAgora.ends } }).then((response) => {
-        this.summerUniversities = response.data.data
+      await this.axios.get(this.services['summeruniversity'] + '/recents', { params: { ends: this.selectedAgora.ends } }).then((summerUniversityResponse) => {
+        this.summerUniversities = summerUniversityResponse.data.data
       }).catch((err) => {
-        this.isLoading = false
         this.$root.showError('Could not fetch Summer University data', err)
       })
     },
@@ -306,7 +312,7 @@ export default {
       const promises = []
       promises.push(this.fetchEvents())
       promises.push(this.fetchStatutoryEvents())
-      promises.push(this.fetchSummerUniversities())
+      // promises.push(this.fetchSummerUniversities())
 
       await Promise.all(promises)
 
@@ -323,17 +329,17 @@ export default {
         body.latest_event = !body.latest_event || moment(event.latest_event).isAfter(moment(body.latest_event)) ? event.latest_event : body.latest_event
       }
 
-      for (const event of this.summerUniversities) {
-        for (const organizer of event.organizing_bodies) {
-          const body = this.bodies.find(x => x.id === organizer.body_id)
-          body.latest_event = !body.latest_event || moment(event.latest_event).isAfter(moment(body.latest_event)) ? event.latest_event : body.latest_event
-        }
-      }
+      // for (const event of this.summerUniversities) {
+      //   for (const organizer of event.organizing_bodies) {
+      //     const body = this.bodies.find(x => x.id === organizer.body_id)
+      //     body.latest_event = !body.latest_event || moment(event.latest_event).isAfter(moment(body.latest_event)) ? event.latest_event : body.latest_event
+      //   }
+      // }
       // ... until here
 
       for (const body in this.bodies) {
         // Check if the last event is in the past 2 years
-        this.bodies[body].antennaCriteria.events = this.bodies[body].latest_event !== undefined && moment(this.bodies[body].latest_event).diff(moment(this.selectedAgora.ends), 'years', true) <= 2
+        this.bodies[body].antennaCriteria.events = String(this.bodies[body].latest_event !== undefined && moment(this.bodies[body].latest_event).diff(moment(this.selectedAgora.ends), 'years', true) <= 2)
         if (this.bodies[body].latest_event !== undefined) this.bodies[body].latest_event = moment(this.bodies[body].latest_event).format('M[/]YYYY')
       }
     },
@@ -346,21 +352,20 @@ export default {
 
         for (const body in this.bodies) {
           // Check if the current board was elected within the past year
-          this.bodies[body].antennaCriteria.boardElection = this.bodies[body].latest_election !== undefined && moment(this.bodies[body].latest_election).diff(moment(this.selectedAgora.ends), 'years', true) <= 1
+          this.bodies[body].antennaCriteria.boardElection = String(this.bodies[body].latest_election !== undefined && moment(this.bodies[body].latest_election).diff(moment(this.selectedAgora.ends), 'years', true) <= 1)
           if (this.bodies[body].latest_election !== undefined) this.bodies[body].latest_election = moment(this.bodies[body].latest_election).format('D[/]M[/]YYYY')
         }
       }).catch((err) => {
-        this.isLoading = false
         this.$root.showError('Could not fetch boards data', err)
       })
     },
     async checkMembersList () {
-      await this.axios.get(this.services['statutory'] + '/events/' + this.selectedAgora.id + '/memberslists/missing').then((membersListResponse) => {
-        for (const body in this.bodies) {
-          this.bodies[body].antennaCriteria.membersList = this.bodies[body].id in membersListResponse.data.data.map(x => x.id)
+      await this.axios.get(this.services['statutory'] + '/events/' + this.selectedAgora.id + '/memberslists').then((membersListsResponse) => {
+        for (const membersList of membersListsResponse.data.data) {
+          const body = this.bodies.find(x => x.id === membersList.body_id)
+          body.antennaCriteria.membersList = 'true'
         }
       }).catch((err) => {
-        this.isLoading = false
         this.$root.showError('Could not fetch members list data', err)
       })
     },
@@ -376,7 +381,6 @@ export default {
           body.comments[criterionName] = criterion.comment
         }
       }).catch((err) => {
-        this.isLoading = false
         this.$root.showError('Could not fetch manual Antenna Criteria fulfilment', err)
       })
     }